+++ src/main/java/es/deusto/sd/auctions/entity/Bid.java
@@ -26,7 +26,7 @@ public class Bid {
     private long date;
 
     @Column(nullable = false)
-    private float amount;
+    private double amount;
 
     @ManyToOne
     @JoinColumn(name = "article_id", nullable = false)
@@ -40,7 +40,7 @@ public class Bid {
 	public Bid() { }
 	
 	// Constructor with parameters
-	public Bid(long date, float amount, Article article, User user) {
+	public Bid(long date, double amount, Article article, User user) {
 		this.date = date;
 		this.amount = amount;
 		this.article = article;
@@ -56,11 +56,11 @@ public class Bid {
 		this.date = date;
 	}
 
-	public float getAmount() {
+	public double getAmount() {
 		return amount;
 	}
 
-	public void setAmount(float amount) {
+	public void setAmount(double amount) {
 		this.amount = amount;
 	}
