+++ src/main/java/es/deusto/sd/auctions/external/ICurrencyServiceGateway.java
@@ -3,5 +3,5 @@ package es.deusto.sd.auctions.external;
 import java.util.Optional;
 
 public interface ICurrencyServiceGateway {
-	public Optional<Float> getExchangeRate(String baseCurrency, String targetCurrency);
+	public Optional<Double> getExchangeRate(String baseCurrency, String targetCurrency);
 }
